refactor: split the selector and reducer and create a variable for initialState
