+++ pointer-poker/src/redux/user/selectors.ts
@@ -1,0 +1,3 @@
+import { RootState } from '../store';
+
+export const selectUser = (state: RootState) => state.user;
